Extract box-related functions to a separate file
